Refactor sample to include on main.dart because they need display on pub.dev site.
